pyproject.toml
[project]
name = "atlas-trader"
version = "0.1.0"
description = "Agentic AI Trading System with Claude API"
requires-python = ">=3.11"
dependencies = [
    # Core
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "python-dotenv>=1.0",
    # Data
    "polars>=1.0",
    "pandas>=2.2",
    "duckdb>=1.0",
    "ccxt>=4.0",
    "yfinance>=0.2",
    "MetaTrader5>=5.0; sys_platform == 'win32'",
    "newsapi-python>=0.2",
    "feedparser>=6.0",
    "aiohttp>=3.9",
    "websockets>=12.0",
    # Analysis
    "pandas-ta>=0.3",
    "numpy>=1.26",
    "scipy>=1.12",
    "hmmlearn>=0.3",
    # AI
    "anthropic>=0.40",
    # Execution & Monitoring
    "apscheduler>=3.10",
    "streamlit>=1.38",
    "plotly>=5.22",
    "python-telegram-bot>=21.0",
    # Testing
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[project.scripts]
atlas-backtest = "scripts.run_backtest:main"
atlas-paper = "scripts.run_paper:main"
atlas-live = "scripts.run_live:main"
atlas-dashboard = "dashboard.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
